Added lima_memtester config
This can be used to build iniramfs image, which automatically
starts the lima-memtester program. If the hardware has LEDs,
then they are used to indicate the current status as it
transitions between the following states:

1. One LED is blinking, turned off most of the time ("A short
   test is currently running").
2. One LED is blinking, equal duration for the on and off
   phases ("A short test has successfully finished and a long
   test is currently running").
3. One LED is blinking, turned on most of the time. And all the
   other LEDs are turned on ("Both short and long tests have
   successfully completed, but the lima-memtester program still
   keeps running"). The test has successfully passed.

In the case if any data curruption is detected, all the LEDs
get turned off. This indicates a test failure. Additionally,
if the device deadlocks, then the status LED may get stuck in
either on or off state. This also indicates a failure.

To sum it up: the device is considered to be still likely alive
if at least one of the LEDs keeps blinking.

And of course, if a monitor or an LCD screen is available, then
the usual spinning cube animation is supposed to be observed.
This LEDs support is intended to be used only when there is no
spare HDMI monitor.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>